+++ pyquil/latex/__init__.py
@@ -1,2 +1,2 @@
 from pyquil.latex.latex_generation import to_latex, DiagramSettings
-from pyquil.latex.ipython import display
+from pyquil.latex._ipython import display
