+++ pyquil/latex/_ipython.py
